+++ game_patch/misc/alpine_settings.cpp
@@ -308,6 +308,18 @@ bool alpine_player_settings_load(rf::Player* player)
     else {
         rf::game_set_gore_level(2); // if gore level not in ini file, default to 2
     }
+    if (settings.count("GibChunkCount")) {
+        g_alpine_game_config.set_gib_chunk_count(std::stoi(settings["GibChunkCount"]));
+        processed_keys.insert("GibChunkCount");
+    }
+    if (settings.count("GibVelocityScale")) {
+        g_alpine_game_config.set_gib_velocity_scale(std::stof(settings["GibVelocityScale"]));
+        processed_keys.insert("GibVelocityScale");
+    }
+    if (settings.count("GibLifetimeMs")) {
+        g_alpine_game_config.set_gib_lifetime_ms(std::stoi(settings["GibLifetimeMs"]));
+        processed_keys.insert("GibLifetimeMs");
+    }
 
     if (settings.count("ShowFPGun")) {
         player->settings.render_fpgun = std::stoi(settings["ShowFPGun"]);
@@ -1056,6 +1068,9 @@ void alpine_player_settings_save(rf::Player* player)
     file << "\n[PlayerSettings]\n";
     file << "PlayerName=" << player->name << "\n";
     file << "GoreLevel=" << rf::game_get_gore_level() << "\n";
+    file << "GibChunkCount=" << g_alpine_game_config.gib_chunk_count << "\n";
+    file << "GibVelocityScale=" << g_alpine_game_config.gib_velocity_scale << "\n";
+    file << "GibLifetimeMs=" << g_alpine_game_config.gib_lifetime_ms << "\n";
     file << "ShowFPGun=" << player->settings.render_fpgun << "\n";
     file << "AutoswitchWeapons=" << player->settings.autoswitch_weapons << "\n";
     file << "NeverAutoswitchExplosives=" << player->settings.dont_autoswitch_to_explosives << "\n";