+++ game_patch/multi/server_internal.h
@@ -519,6 +519,20 @@ struct ClickLimiterConfig
     }
 };
 
+struct GibConfig
+{
+    bool enabled = true;
+    bool all_damage = false;
+    float damage_threshold = 100.0f;
+
+    // =============================================
+
+    void set_damage_threshold(float threshold)
+    {
+        damage_threshold = std::clamp(threshold, 1.0f, 6000.0f);
+    }
+};
+
 struct AlpineServerConfigRules
 {
     // stock game rules
@@ -536,7 +550,6 @@ struct AlpineServerConfigRules
     bool weapons_stay = false;
     bool force_respawn = false;
     bool balance_teams = false;
-    bool gib_explosive_deaths = true;
     int ideal_player_count = 32;
     bool saving_enabled = false;
     bool flag_dropping = true;
@@ -553,6 +566,7 @@ struct AlpineServerConfigRules
     WeaponLoadoutConfig spawn_loadout;
     SpawnProtectionConfig spawn_protection;
     NewSpawnLogicConfig spawn_logic;
+    GibConfig gibbing;
     WelcomeMessageConfig welcome_message;
     bool weapon_items_give_full_ammo = false;
     bool weapon_infinite_magazines = false;